# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface ResearchServiceStatus {
  aiAvailable: boolean;
  aiProvider: string;
  searchAvailable: boolean;
  searchProvider: string;
}

export interface ResearchSearchInput {
  /**
     * @minLength 3
     * @maxLength 500
     */
  question: string;
  /**
     * @minimum 5
     * @maximum 15
     */
  limit: number;
}

export type ResearchSourceCredibility = typeof ResearchSourceCredibility[keyof typeof ResearchSourceCredibility];


export const ResearchSourceCredibility = {
  high: 'high',
  medium: 'medium',
} as const;

export interface ResearchSource {
  id: string;
  title: string;
  organization: string;
  url: string;
  /** @nullable */
  author: string | null;
  /** @nullable */
  publishedDate: string | null;
  /** @nullable */
  abstract: string | null;
  sourceType: string;
  /** @nullable */
  doi: string | null;
  credibility: ResearchSourceCredibility;
}

export interface ResearchSearchResponse {
  sources: ResearchSource[];
  providers: string[];
  searchedAt: string;
  /** @nullable */
  warning: string | null;
}