# lib/api-zod/src/generated/types/researchSearchInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResearchSearchInput {
  /**
     * @minLength 3
     * @maxLength 500
     */
  question: string;
  /**
     * @minimum 5
     * @maximum 15
     */
  limit: number;
}