# lib/api-zod/src/generated/types/researchSearchResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResearchSource } from './researchSource';

export interface ResearchSearchResponse {
  sources: ResearchSource[];
  providers: string[];
  searchedAt: Date;
  /** @nullable */
  warning: string | null;
}