# lib/api-zod/src/generated/types/researchServiceStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResearchServiceStatus {
  aiAvailable: boolean;
  aiProvider: string;
  searchAvailable: boolean;
  searchProvider: string;
}