# lib/api-zod/src/generated/types/researchSource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResearchSourceCredibility } from './researchSourceCredibility';

export interface ResearchSource {
  id: string;
  title: string;
  organization: string;
  url: string;
  /** @nullable */
  author: string | null;
  /** @nullable */
  publishedDate: string | null;
  /** @nullable */
  abstract: string | null;
  sourceType: string;
  /** @nullable */
  doi: string | null;
  credibility: ResearchSourceCredibility;
}